Separate HEKI/HVCI from `litebox_platform_lvbs` (#1093)
This PR separates the HEKI/HVCI features from `litebox_platform_lvbs`.
It introduces a `litebox_service_heki` crate which hosts all HEKI/HVCI
features (algorithms). This crate and the runner rely on `Vtl0Gate`,
`Vtl0PrivilegedWrite`, and `Vtl1Gate` abstraction concretized by
`litebox_platform_lvbs` to gate VTL0 interaction and self-protect VTL1.

---------
