+++ litebox_runner_lvbs/Cargo.toml
@@ -7,10 +7,11 @@ edition = "2024"
 arrayvec = { version = "0.7.6", default-features = false }
 litebox = { version = "0.1.0", path = "../litebox" }
 litebox_platform_lvbs = { version = "0.1.0", path = "../litebox_platform_lvbs", default-features = false }
-litebox_common_lvbs = { version = "0.1.0", path = "../litebox_common_lvbs" }
 litebox_platform_multiplex = { version = "0.1.0", path = "../litebox_platform_multiplex", default-features = false, features = ["platform_lvbs"] }
 litebox_common_optee = { path = "../litebox_common_optee/", version = "0.1.0" }
 litebox_common_linux = { path = "../litebox_common_linux/", version = "0.1.0" }
+litebox_common_lvbs = { path = "../litebox_common_lvbs/", version = "0.1.0" }
+litebox_service_heki = { path = "../litebox_service_heki/", version = "0.1.0" }
 litebox_shim_optee = { path = "../litebox_shim_optee/", version = "0.1.0" }
 litebox_util_log = { version = "0.1.0", path = "../litebox_util_log" }
 log = { version = "0.4", default-features = false }
